requirements.txt
fastapi
uvicorn[standard]
pillow
numpy
ultralytics
python-multipart
aiofiles
opencv-python-headless
